Added HKL matching and `pull_spots()` comparison
Everything seems to be working pretty well for the most part. We are
finding 99.8% of all the spots that `pull_spots()` found. They are
somewhat close as well in both `x, y` and `omega` compared to the
output of `pull_spots()`.

Some extra spots were also assigned to HKLs, and those are worth
further investigation.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>